Split SessionMaster into its own file from LiftSession.
